@@ -1,13 +1,18 @@
-export const dirname = (path) => {
+export default function dirname(path) {
   if (typeof path !== 'string') {
-    throw new TypeError('Path must be a string. Received ' + JSON.stringify(path));
+    throw new TypeError(`Path must be a string. Received ${JSON.stringify(path)}`);
   }
 
-  if (path.length === 0) return '.';
+  if (path.length === 0) {
+    return '.';
+  }
+
   let code = path.charCodeAt(0);
   const hasRoot = code === 47;
   let end = -1;
   let matchedSlash = true;
+
+  // eslint-disable-next-line no-plusplus
   for (let i = path.length - 1; i >= 1; --i) {
     code = path.charCodeAt(i);
     if (code === 47) {
@@ -16,12 +21,17 @@ export const dirname = (path) => {
         break;
       }
     } else {
-      // We saw the first non-path separator
       matchedSlash = false;
     }
   }
 
-  if (end === -1) return hasRoot ? '/' : '.';
-  if (hasRoot && end === 1) return '//';
+  if (end === -1) {
+    return hasRoot ? '/' : '.';
+  }
+
+  if (hasRoot && end === 1) {
+    return '//';
+  }
+
   return path.slice(0, end);
 }
