@@ -5,13 +5,13 @@
   >
     <ol>
       <li
-        v-for="(val, index) in crumbs"
+        v-for="(val, index) in crumbItems()"
         :key="index"
         class="media-breadcrumb-item"
       >
         <a
           href="#"
-          :aria-current="(index === Object.keys(crumbs).length - 1) ? 'page' : undefined"
+          :aria-current="(index === Object.keys(crumbItems()).length - 1) ? 'page' : undefined"
           @click.stop.prevent="onCrumbClick(index)"
         >{{ val.name }}</a>
       </li>
@@ -19,83 +19,83 @@
   </nav>
 </template>
 
-<script>
-import navigable from '../../mixins/navigable.es6';
+<script setup>
+import { computed } from 'vue';
+import { useFileStore } from '../../stores/files.es6.js';
 
-export default {
-  name: 'MediaBreadcrumb',
-  mixins: [navigable],
-  computed: {
-    /* Get the crumbs from the current directory path */
-    crumbs() {
-      const items = [];
-      const adapter = this.$store.state.selectedDirectory.split(':/');
+const fileStore = useFileStore();
+const disks = computed(() => fileStore.disks);
+const selectedDirectory = computed(() => fileStore.selectedDirectory);
 
-      // Add the drive as first element
-      if (adapter.length) {
-        const drive = this.findDrive(adapter[0]);
+/* Handle the on crumb click event */
+function onCrumbClick(index) {
+  const destination = crumbItems().find((crumb) => crumb.index === index);
 
-        if (!drive) {
-          return [];
-        }
+  if (!destination) {
+    return;
+  }
 
-        items.push(drive);
-        let path = `${adapter[0]}:`;
+  fileStore.getPathContents(destination.path, false, false);
+  window.parent.document.dispatchEvent(
+    new CustomEvent(
+      'onMediaFileSelected',
+      {
+        bubbles: true,
+        cancelable: false,
+        detail: {},
+      },
+    ),
+  );
+}
 
-        adapter[1]
-          .split('/')
-          .filter((crumb) => crumb.length !== 0)
-          .forEach((crumb, index) => {
-            path = `${path}/${crumb}`;
-            items.push({
-              name: crumb,
-              index: index + 1,
-              path,
-            });
-          });
+function findDrive(adapter) {
+  let driveObject = null;
+
+  disks.value.forEach((disk) => {
+    disk.drives.forEach((drive) => {
+      if (drive.root.startsWith(adapter)) {
+        driveObject = { name: drive.displayName, path: drive.root, index: 0 };
       }
+    });
+  });
 
-      return items;
-    },
-    /* Whether or not the crumb is the last element in the list */
-    isLast(item) {
-      return this.crumbs.indexOf(item) === this.crumbs.length - 1;
-    },
-  },
-  methods: {
-    /* Handle the on crumb click event */
-    onCrumbClick(index) {
-      const destination = this.crumbs.find((crumb) => crumb.index === index);
+  return driveObject;
+}
 
-      if (!destination) {
-        return;
-      }
+/* Get the crumbs from the current directory path */
+function crumbItems() {
+  const items = [];
+  const adapter = selectedDirectory.value.split(':/');
 
-      this.navigateTo(destination.path);
-      window.parent.document.dispatchEvent(
-        new CustomEvent(
-          'onMediaFileSelected',
-          {
-            bubbles: true,
-            cancelable: false,
-            detail: {},
-          },
-        ),
-      );
-    },
-    findDrive(adapter) {
-      let driveObject = null;
+  // Add the drive as first element
+  if (adapter.length) {
+    const drive = findDrive(adapter[0]);
 
-      this.$store.state.disks.forEach((disk) => {
-        disk.drives.forEach((drive) => {
-          if (drive.root.startsWith(adapter)) {
-            driveObject = { name: drive.displayName, path: drive.root, index: 0 };
-          }
+    if (!drive) {
+      return [];
+    }
+
+    items.push(drive);
+    let path = `${adapter[0]}:`;
+
+    adapter[1]
+      .split('/')
+      .filter((crumb) => crumb.length !== 0)
+      .forEach((crumb, index) => {
+        path = `${path}/${crumb}`;
+        items.push({
+          name: crumb,
+          index: index + 1,
+          path,
         });
       });
+  }
 
-      return driveObject;
-    },
-  },
-};
+  return items;
+}
+
+/* Whether or not the crumb is the last element in the list */
+function isLast(item) {
+  return crumbItems().indexOf(item) === crumbItems().length - 1;
+}
 </script>
