@@ -2,7 +2,7 @@
   <tr
     class="media-browser-item"
     :class="{selected}"
-    @dblclick.stop.prevent="onDblClick()"
+    @dblclick.stop.prevent="onDblClick"
     @click="onClick"
   >
     <td
@@ -31,17 +31,57 @@
 </template>
 
 <script>
-import * as types from '../../../store/mutation-types.es6';
-import navigable from '../../../mixins/navigable.es6';
+import { computed } from 'vue';
+import { useFileStore } from '../../../stores/files.es6.js';
+import { useViewStore } from '../../../stores/listview.es6.js';
 
 export default {
   name: 'MediaBrowserItemRow',
-  mixins: [navigable],
   props: {
     item: {
       type: Object,
       default: () => {},
     },
+  },
+  setup() {
+    const fileStore = useFileStore();
+    const disks = computed(() => fileStore.disks);
+    const directories = computed(() => fileStore.directories);
+    const selectedDirectory = computed(() => fileStore.selectedDirectory);
+    const selectedItems = computed(() => fileStore.selectedItems);
+    const search = computed(() => fileStore.search);
+
+    const viewStore = useViewStore();
+    const loading = computed(() => viewStore.loading);
+    const showInfoBar = computed(() => viewStore.showInfoBar);
+    const listView = computed(() => viewStore.listView);
+    const gridSize = computed(() => viewStore.gridSize);
+    const previewItem = computed(() => viewStore.previewItem);
+    const sortBy = computed(() => viewStore.sortBy);
+    const sortDirection = computed(() => viewStore.sortDirection);
+
+    return {
+      // disks,
+      // directories,
+      // selectedDirectory,
+      selectedItems,
+      // search,
+      fileStore,
+
+      // loading,
+      // showInfoBar,
+      // listView,
+      // gridSize,
+      // showConfirmDeleteModal,
+      // showCreateFolderModal,
+      // showPreviewModal,
+      // showShareModal,
+      // showRenameModal,
+      // previewItem,
+      // sortBy,
+      // sortDirection,
+      // viewStore,
+    };
   },
   computed: {
     /* The dimension of a file */
@@ -70,7 +110,7 @@ export default {
     /* Handle the on row double click event */
     onDblClick() {
       if (this.isDir) {
-        this.navigateTo(this.item.path);
+        this.fileStore.getPathContents(this.item.path, false, false);
         return;
       }
 
@@ -80,8 +120,8 @@ export default {
       // Show preview
       if (this.item.extension
         && extensionWithPreview.includes(this.item.extension.toLowerCase())) {
-        this.$store.commit(types.SHOW_PREVIEW_MODAL);
-        this.$store.dispatch('getFullContents', this.item);
+        this.viewStore.showPreviewModal();
+        this.fileStore.getFullContents(this.item);
       }
     },
 
@@ -90,7 +130,7 @@ export default {
      * @returns {boolean}
      */
     isSelected() {
-      return this.$store.state.selectedItems.some((selected) => selected.path === this.item.path);
+      return this.selectedItems.some((selected) => selected.path === this.item.path);
     },
 
     /**
@@ -129,17 +169,18 @@ export default {
         // Unselect all other selected items,
         // if the shift key was not pressed during the click event
         if (!(event.shiftKey || event.keyCode === 13)) {
-          this.$store.commit(types.UNSELECT_ALL_BROWSER_ITEMS);
+          this.fileStore.resetSelectedItems();
         }
-        this.$store.commit(types.SELECT_BROWSER_ITEM, this.item);
+        this.fileStore.addItemToSelectedItems(this.item);
         return;
       }
 
       // If more than one item was selected and the user clicks again on the selected item,
       // he most probably wants to unselect all other items.
-      if (this.$store.state.selectedItems.length > 1) {
-        this.$store.commit(types.UNSELECT_ALL_BROWSER_ITEMS);
-        this.$store.commit(types.SELECT_BROWSER_ITEM, this.item);
+      if (this.selectedItems.length > 1) {
+        console.log('nope')
+        this.fileStore.resetSelectedItems();
+        this.fileStore.addItemToSelectedItems(this.item);
       }
     },
 
