@@ -1,6 +1,6 @@
 <template>
   <MediaModal
-    v-if="$store.state.showCreateFolderModal"
+    v-if="showCreateFolderModal"
     :size="'md'"
     label-element="createFolderTitle"
     @close="close()"
@@ -57,13 +57,62 @@
 </template>
 
 <script>
-import * as types from '../../store/mutation-types.es6';
+import {
+  computed, defineComponent, onMounted, ref,
+} from 'vue';
 import MediaModal from './modal.vue';
+import { useFileStore } from '../../stores/files.es6.js';
+import { useViewStore } from '../../stores/listview.es6.js';
 
 export default {
   name: 'MediaCreateFolderModal',
   components: {
     MediaModal,
+  },
+  setup() {
+    const filesStore = useFileStore();
+    const disks = computed(() => filesStore.disks);
+    const directories = computed(() => filesStore.directories);
+    const selectedDirectory = computed(() => filesStore.selectedDirectory);
+    const selectedItems = computed(() => filesStore.selectedItems);
+    const search = computed(() => filesStore.search);
+
+    const viewStore = useViewStore();
+    const isLoading = computed(() => filesStore.isLoading);
+    const showInfoBar = computed(() => filesStore.showInfoBar);
+    const listView = computed(() => filesStore.listView);
+    const gridSize = computed(() => filesStore.gridSize);
+    const showConfirmDeleteModal = computed(() => filesStore.showConfirmDeleteModal);
+    const showCreateFolderModal = computed(() => filesStore.showCreateFolderModal);
+    const showPreviewModal = computed(() => filesStore.showPreviewModal);
+    const showShareModal = computed(() => filesStore.showShareModal);
+    const showRenameModal = computed(() => filesStore.showRenameModal);
+    const previewItem = computed(() => filesStore.previewItem);
+    const sortBy = computed(() => filesStore.sortBy);
+    const sortDirection = computed(() => filesStore.sortDirection);
+
+    return {
+      disks,
+      directories,
+      selectedDirectory,
+      selectedItems,
+      search,
+      filesStore,
+
+      isLoading,
+      showInfoBar,
+      listView,
+      gridSize,
+      showConfirmDeleteModal,
+      showCreateFolderModal,
+      showPreviewModal,
+      showShareModal,
+      showRenameModal,
+      previewItem,
+      sortBy,
+      sortDirection,
+      viewStore,
+    };
   },
   data() {
     return {
@@ -71,7 +120,7 @@ export default {
     };
   },
   watch: {
-    '$store.state.showCreateFolderModal': function (show) {
+    'store.showCreateFolderModal': function (show) {
       this.$nextTick(() => {
         if (show && this.$refs.input) {
           this.$refs.input.focus();
@@ -87,7 +136,7 @@ export default {
     /* Close the modal instance */
     close() {
       this.reset();
-      this.$store.commit(types.HIDE_CREATE_FOLDER_MODAL);
+      this.store.hideCreateFolderModal();
     },
     /* Save the form and create the folder */
     save() {
@@ -99,9 +148,9 @@ export default {
       }
 
       // Create the directory
-      this.$store.dispatch('createDirectory', {
+      this.store.createDirectory({
         name: this.folder,
-        parent: this.$store.state.selectedDirectory,
+        parent: this.store.selectedDirectory(),
       });
       this.reset();
     },
