@@ -1,6 +1,6 @@
 <template>
   <MediaModal
-    v-if="$store.state.showPreviewModal && item"
+    v-if="showPreviewModal && item"
     :size="'md'"
     class="media-preview-modal"
     label-element="previewTitle"
@@ -59,20 +59,69 @@
 </template>
 
 <script>
+import {
+  computed, watch, onMounted, ref,
+} from 'vue';
 import api from '../../app/Api.es6';
-import * as types from '../../store/mutation-types.es6';
 import MediaModal from './modal.vue';
+import { useFileStore } from '../../stores/files.es6.js';
+import { useViewStore } from '../../stores/listview.es6.js';
 
 export default {
   name: 'MediaPreviewModal',
   components: {
     MediaModal,
   },
+  setup() {
+    const filesStore = useFileStore();
+    const disks = computed(() => filesStore.disks);
+    const directories = computed(() => filesStore.directories);
+    const selectedDirectory = computed(() => filesStore.selectedDirectory);
+    const selectedItems = computed(() => filesStore.selectedItems);
+    const search = computed(() => filesStore.search);
+
+    const viewStore = useViewStore();
+    const isLoading = computed(() => filesStore.isLoading);
+    const showInfoBar = computed(() => filesStore.showInfoBar);
+    const listView = computed(() => filesStore.listView);
+    const gridSize = computed(() => filesStore.gridSize);
+    const showConfirmDeleteModal = computed(() => filesStore.showConfirmDeleteModal);
+    const showCreateFolderModal = computed(() => filesStore.showCreateFolderModal);
+    const showPreviewModal = computed(() => filesStore.showPreviewModal);
+    const showShareModal = computed(() => filesStore.showShareModal);
+    const showRenameModal = computed(() => filesStore.showRenameModal);
+    const previewItem = computed(() => filesStore.previewItem);
+    const sortBy = computed(() => filesStore.sortBy);
+    const sortDirection = computed(() => filesStore.sortDirection);
+
+    return {
+      disks,
+      directories,
+      selectedDirectory,
+      selectedItems,
+      search,
+      filesStore,
+
+      isLoading,
+      showInfoBar,
+      listView,
+      gridSize,
+      showConfirmDeleteModal,
+      showCreateFolderModal,
+      showPreviewModal,
+      showShareModal,
+      showRenameModal,
+      previewItem,
+      sortBy,
+      sortDirection,
+      viewStore,
+    };
+  },
   computed: {
     /* Get the item to show in the modal */
     item() {
       // Use the currently selected directory as a fallback
-      return this.$store.state.selectedItem ? this.$store.state.selectedItem : this.$store.state.previewItem;
+      return this.store.selectedItem ? this.store.selectedItem : this.store.previewItem;
     },
     /* Get the hashed URL */
     getHashedURL() {
@@ -88,7 +137,7 @@ export default {
   methods: {
     /* Close the modal */
     close() {
-      this.$store.commit(types.HIDE_PREVIEW_MODAL);
+      this.store.hidePreviewModal();
     },
     isImage() {
       return this.item.mime_type.indexOf('image/') === 0;
