@@ -1,6 +1,6 @@
 <template>
   <MediaModal
-    v-if="$store.state.showShareModal"
+    v-if="showShareModal"
     :size="'md'"
     :show-close="false"
     label-element="shareTitle"
@@ -74,34 +74,81 @@
 </template>
 
 <script>
-import * as types from '../../store/mutation-types.es6';
+import { computed, ref } from 'vue';
 import translate from '../../plugins/translate.es6';
 import MediaModal from './modal.vue';
+import { useFileStore } from '../../stores/files.es6.js';
+import { useViewStore } from '../../stores/listview.es6.js';
 
 export default {
   name: 'MediaShareModal',
   components: {
     MediaModal,
   },
+  setup() {
+    const filesStore = useFileStore();
+    const disks = computed(() => filesStore.disks);
+    const directories = computed(() => filesStore.directories);
+    const selectedDirectory = computed(() => filesStore.selectedDirectory);
+    const selectedItems = computed(() => filesStore.selectedItems);
+    const search = computed(() => filesStore.search);
+
+    const viewStore = useViewStore();
+    const isLoading = computed(() => filesStore.isLoading);
+    const showInfoBar = computed(() => filesStore.showInfoBar);
+    const listView = computed(() => filesStore.listView);
+    const gridSize = computed(() => filesStore.gridSize);
+    const showConfirmDeleteModal = computed(() => filesStore.showConfirmDeleteModal);
+    const showCreateFolderModal = computed(() => filesStore.showCreateFolderModal);
+    const showPreviewModal = computed(() => filesStore.showPreviewModal);
+    const showShareModal = computed(() => filesStore.showShareModal);
+    const showRenameModal = computed(() => filesStore.showRenameModal);
+    const previewItem = computed(() => filesStore.previewItem);
+    const sortBy = computed(() => filesStore.sortBy);
+    const sortDirection = computed(() => filesStore.sortDirection);
+
+    return {
+      disks,
+      directories,
+      selectedDirectory,
+      selectedItems,
+      search,
+      filesStore,
+
+      isLoading,
+      showInfoBar,
+      listView,
+      gridSize,
+      showConfirmDeleteModal,
+      showCreateFolderModal,
+      showPreviewModal,
+      showShareModal,
+      showRenameModal,
+      previewItem,
+      sortBy,
+      sortDirection,
+      viewStore,
+    };
+  },
   computed: {
     item() {
-      return this.$store.state.selectedItems[this.$store.state.selectedItems.length - 1];
+      return this.selectedItems[this.store.selectedItems.length - 1];
     },
 
     url() {
-      return (this.$store.state.previewItem && Object.prototype.hasOwnProperty.call(this.$store.state.previewItem, 'url') ? this.$store.state.previewItem.url : null);
+      return (this.viewStore.previewItem && Object.prototype.hasOwnProperty.call(this.viewStore.previewItem, 'url') ? this.viewStore.previewItem.url : null);
     },
   },
   methods: {
     /* Close the modal instance and reset the form */
     close() {
-      this.$store.commit(types.HIDE_SHARE_MODAL);
-      this.$store.commit(types.LOAD_FULL_CONTENTS_SUCCESS, null);
+      this.store.hideShareModal();
+      this.store.loadFullContentsSuccess(null);
     },
 
     // Generate the url from backend
     generateUrl() {
-      this.$store.dispatch('getFullContents', this.item);
+      this.store.getFullContents(this.item);
     },
 
     // Copy to clipboard
