@@ -1,8 +1,8 @@
 import { createApp } from 'vue';
+import { createPinia } from 'pinia';
 import App from './components/app.vue';
-import Event from './app/Event.es6';
-import store from './store/store.es6';
-import translate from './plugins/translate.es6';
+import Event from './app/Event.es6.js';
+import translate from './plugins/translate.es6.js';
 
 // Register MediaManager namespace
 window.MediaManager = window.MediaManager || {};
@@ -10,4 +10,4 @@ window.MediaManager = window.MediaManager || {};
 window.MediaManager.Event = new Event();
 
 // Create the Vue app instance
-createApp(App).use(store).use(translate).mount('#com-media');
+createApp(App).use(createPinia()).use(translate).mount('#com-media');
