@@ -1,7 +1,6 @@
-import persistedStateOptions from './plugins/persisted-state.es6';
-
 // Get the disks from joomla option storage
 const options = Joomla.getOptions('com_media', {});
+
 if (options.providers === undefined || options.providers.length === 0) {
   throw new TypeError('Media providers are not defined.');
 }
@@ -28,14 +27,8 @@ if (!defaultDisk) {
   throw new TypeError('No default media drive was found');
 }
 
-const storedState = JSON.parse(persistedStateOptions.storage.getItem(persistedStateOptions.key));
-
-function setSession(path) {
-  persistedStateOptions.storage.setItem(
-    persistedStateOptions.key,
-    JSON.stringify({ ...storedState, ...{ selectedDirectory: path } }),
-  );
-}
+export const storedState = JSON.parse(window.sessionStorage.getItem('joomla.mediamanager'));
+export const setSession = (path) => window.sessionStorage.setItem('joomla.mediamanager', JSON.stringify({ ...storedState, ...{ selectedDirectory: path } }));
 
 // Gracefully use the given path, the session storage state or fall back to sensible default
 function getCurrentPath() {
@@ -74,49 +67,8 @@ function getCurrentPath() {
   return path;
 }
 
-// The initial state
-export default {
-  // The general loading state
-  isLoading: false,
-  // Will hold the activated filesystem disks
-  disks: loadedDisks,
-  // The loaded directories
-  directories: loadedDisks.map(() => ({
-    path: defaultDisk.drives[0].root,
-    name: defaultDisk.displayName,
-    directories: [],
-    files: [],
-    directory: null,
-  })),
-  // The loaded files
-  files: [],
-  // The selected disk. Providers are ordered by plugin ordering, so we set the first provider
-  // in the list as the default provider and load first drive on it as default
-  selectedDirectory: getCurrentPath(),
-  // The currently selected items
-  selectedItems: [],
-  // The state of the infobar
-  showInfoBar: false,
-  // List view
-  listView: 'grid',
-  // The size of the grid items
-  gridSize: 'md',
-  // The state of confirm delete model
-  showConfirmDeleteModal: false,
-  // The state of create folder model
-  showCreateFolderModal: false,
-  // The state of preview model
-  showPreviewModal: false,
-  // The state of share model
-  showShareModal: false,
-  // The state of  model
-  showRenameModal: false,
-  // The preview item
-  previewItem: null,
-  // The Search Query
-  search: '',
-  // The sorting by
-  sortBy: storedState && storedState.sortBy ? storedState.sortBy : 'name',
-  // The sorting direction
-  sortDirection: storedState && storedState.sortDirection ? storedState.sortDirection : 'asc',
+export {
+  loadedDisks,
+  defaultDisk,
+  getCurrentPath,
 };
